Merge branch 'master' into jgm/daos-12412

Quick-Functional: true
Test-tag: daos_core_test_dfs
Test-provider: ucx+dc_x
PR-repos: mercury@PR-98:2

+++ src/object/cli_ec.c
@@ -1403,7 +1403,7 @@ obj_ec_fail_info_insert(struct obj_reasb_req *reasb_req, uint16_t fail_tgt)
 }
 
 int
-obj_ec_singv_split(daos_unit_oid_t oid, uint32_t layout_ver, struct daos_oclass_attr *oca,
+obj_ec_singv_split(daos_unit_oid_t oid, uint16_t layout_ver, struct daos_oclass_attr *oca,
 		   uint64_t dkey_hash, daos_size_t iod_size, d_sg_list_t *sgl)
 {
 	uint64_t c_bytes = obj_ec_singv_cell_bytes(iod_size, oca);
@@ -1442,7 +1442,7 @@ out:
 }
 
 int
-obj_ec_singv_encode_buf(daos_unit_oid_t oid, uint32_t layout_ver, struct daos_oclass_attr *oca,
+obj_ec_singv_encode_buf(daos_unit_oid_t oid, uint16_t layout_ver, struct daos_oclass_attr *oca,
 			uint64_t dkey_hash, daos_iod_t *iod, d_sg_list_t *sgl,
 			d_iov_t *e_iov)
 {
